The code of Board has a new structure and passes the current testsuite.
There are performance and logic problems however. The legality for one field
is checked more than once. The next step is to enhance the testsuite to cover
this error and fix it. Then improve the performance - the new code should run
about as fast (if not faster) than the old one.